+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Reportnet 3 changes
 
+## v1.1.0
+* "Bulk Size" parameter that controls how many FME Features is now a writer parameter instead of a feature type parameter
+* Writer parameter "Connection Timeout" was added
+
 ## v1.0.1
 README updated